% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.r
\name{factor2numeric}
\alias{factor2numeric}
\title{Factor Object To Numeric Object Conversion}
\usage{
factor2numeric(vec)
}
\arguments{
\item{vec}{A vector of elements, typically with class factor}
}
\description{
This function converts categorical vectors to numeric vectors with the
lowest category set to 0. If a factor is passed to obsval() as the
effect.var, this function is used to convert it to numeric values.  It
returns both the converted vector and a key-value list that maps factor
levels to their corresponding numeric values. See example below for more.
}
\examples{
# create factor vector
gender <- factor(c("male","female","male","male"))

# run factor2numeric
convert <- factor2numeric( gender )

# retrieve numeric vector
(gender <- convert$numvec)

# retrieve numeric values corresponding to initial factor levels
(male.val <- convert$keys[["male"]])
(female.val <- convert$keys[["female"]])
}

