% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_range.R
\name{graph_range}
\alias{graph_range}
\title{Graph Effect Range}
\usage{
graph_range(obsval_obj, dv_category = NULL)
}
\arguments{
\item{obsval_obj}{The results from obsval() or obsvalPredict().}

\item{dv_category}{If the dependent variable is categorical, this specifies which category's probabilities to graph.}
}
\description{
This function generates simulated coefficients based on coefficient means and covariances.
}
\examples{
library(mvtnorm)
library(MASS)
library(ggplot2)


data       	<- quine
specification 		<- 'Days~Eth+Sex+Age'
reg.model 			<- "negbin"
n.draws				<- 100
effect.var			<- "Age"
effect.var.high		<- "F3"
effect.var.low		<- "F0"

mymodel <- obsval( data, specification, reg.model, n.draws,
                   effect.var, effect.var.low,effect.var.high,
                   options=list(verbose=TRUE,
                                 effect_range="full"))

graph_range( mymodel )
}

