% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obsval.R
\name{obsval}
\alias{obsval}
\title{Obsval: Using Observed Values to Calculate Predicted Effects}
\usage{
obsval(fmla, data, reg.model = NULL, n.draws = 1000, effect.var = NULL,
  effect.vals = NULL, cluster.se = NULL, subsample.var = NULL,
  subsample.val = NULL, baseline.category = NULL, verbose = FALSE,
  seed = 123, ci = 95)
}
\arguments{
\item{fmla}{A string or formula containing the model specification. Should be of the form 'x ~ y'.}

\item{data}{Data frame containing variables specified by effect.var and fmla.}

\item{reg.model}{A string containing the type of regression to run. Available options are ols, logit, ologit, mlogit, probit, oprobit, poisson, negbin and mlogit.}

\item{n.draws}{An integer specifying how many draws of coefficient estimates to make. Defaults to 1000.}

\item{effect.var}{A string containing the name of the variable whose values will be specified in effect.var.low and effect.var.high}

\item{effect.vals}{An integer or vector specifying the values of effect.var to use. Values can be numeric or character. Effect.preds are calculated by substracting the predictions for final value of effect.vals from the predictions for the initial value.}

\item{cluster.se}{A string specifying the variable name to use as the cluster variable when computing cluster-robust standard errors.}

\item{subsample.var}{A string specifying the variable to use to subset the data post-estimation.}

\item{subsample.val}{A string or integer specifying the value to use to subset the data post-estimation. If subsample.var is "Sex" and subsample.val is "Male", then only observations in which the "Sex" variable is set to "Male" will be used to compute effects.}

\item{baseline.category}{A string specifying the category of the dependent variable to use as a baseline when estimating multinomial models.}

\item{verbose}{A boolean specifying whether to print messages when running obsval(). Useful for debugging. Defaults to FALSE. intervoptions A list of optional variables, including verbose, baseline.category, seed and sample_var.}

\item{seed}{An integer specifying the seed to use when computing simulations. Defaults to 123.}

\item{ci}{An integer specifying the confidence interval to use when returning effect summary tables. Defaults to 95.}
}
\description{
This function calculates predicted probabilities, discrete differences and marginal effects using the observed values of covariates rather than holding covariates constant at their means. It is based on Hanmer and Kalkan (2013).
}
\examples{
# run the model + get estimates
mymodel <- obsval(binary_warsup~female+partyid,
                   data = studentVote,
                   reg.model = "logit",
                   n.draws = 100,
                   effect.var = "female",
                   effect.vals = c(0,1),
                   verbose = TRUE)

# get mean and quantile estimates
mean(mymodel$effect_preds)
quantile(mymodel$effect_preds, c(0.025,0.975))

# visualize estimated effect on probability
#boxplot(mymodel$effect_preds)
}

